Add files via upload
The source code is all in the .py file if you want to alter settings. 

If you just want something to work use the exe file. 

This is not the best code I have wrote and could be cleaned up significantly. I wrote this in one week to serve a business purpose and it has not let us down since. If you find any bugs please reach out to me I will do my best to maintain the project